floating point calculations

+++ main.cpp
@@ -8,7 +8,7 @@ using std::chrono::high_resolution_clock;
 
 #include <glm/glm.hpp>
 #include <glm/gtx/rotate_vector.hpp>
-using glm::dvec3;
+using glm::vec3;
 using glm::dvec2;
 using glm::dmat3;
 
@@ -57,42 +57,42 @@ int main(int argc, char** argv) {
 	// eventually would get replaced with BVH with unique ptr
 	Scene cornellBox({
 		// smallpt scene except light
-		new Sphere(1e5,  dvec3( 1e5+1,40.8,81.6),       dvec3(.75,.25,.25),  dvec3(0.0)),//Left 
-		new Sphere(1e5,  dvec3(-1e5+99,40.8,81.6),      dvec3(.25,.25,.75),  dvec3(0.0)),//Right 
-		new Sphere(1e5,  dvec3( 50.,40.8, 1e5),         dvec3(.75,.75,.75),  dvec3(0.0)),//Back 
-		new Sphere(1e5,  dvec3( 50.,40.8,-1e5+170),     dvec3(0.0),          dvec3(0.0)),//Front 
-		new Sphere(1e5,  dvec3( 50., 1e5, 81.6),        dvec3(.75,.75,.75),  dvec3(0.0)),//Bottom 
-		new Sphere(1e5,  dvec3( 50.,-1e5+81.6,81.6),    dvec3(.75,.75,.75),  dvec3(0.0)),//Top 
-		new Sphere(16.5, dvec3( 27.,16.5,47),           dvec3(1.0)*.999,     dvec3(0.0)),//Mirror 
-		new Sphere(16.5, dvec3( 73.,16.5,78),           dvec3(1.0)*.999,     dvec3(0.0)),//Glass
+		new Sphere(1e5,  vec3( 1e5+1,40.8,81.6),       vec3(.75,.25,.25),  vec3(0.0)),//Left 
+		new Sphere(1e5,  vec3(-1e5+99,40.8,81.6),      vec3(.25,.25,.75),  vec3(0.0)),//Right 
+		new Sphere(1e5,  vec3( 50.,40.8, 1e5),         vec3(.75,.75,.75),  vec3(0.0)),//Back 
+		new Sphere(1e5,  vec3( 50.,40.8,-1e5+170),     vec3(0.0),          vec3(0.0)),//Front 
+		new Sphere(1e5,  vec3( 50., 1e5, 81.6),        vec3(.75,.75,.75),  vec3(0.0)),//Bottom 
+		new Sphere(1e5,  vec3( 50.,-1e5+81.6,81.6),    vec3(.75,.75,.75),  vec3(0.0)),//Top 
+		new Sphere(16.5, vec3( 27.,16.5,47),           vec3(1.0)*.999,     vec3(0.0)),//Mirror 
+		new Sphere(16.5, vec3( 73.,16.5,78),           vec3(1.0)*.999,     vec3(0.0)),//Glass
 	},
 	{
-		/*new Sphere(600., dvec3( 50.,681.6-.27,81.6), dvec3(1.0),dvec3(12.)) //Light */
-		/*new Sphere(5.,dvec3( 50.,81.6-6.,81.6), dvec3(1.0), dvec3(12.)), //Light*/
-		/* Sphere(20.,dvec3( 50.,51.6-6.,81.6), dvec3(1.0), dvec3(12.)) //Light */
-		new Triangle(dvec3( 50.,81.6-9.,81.6),dvec3( 42.,81.6-9.,76.6),dvec3( 42.,81.6-9.,81.6), dvec3(1.0), dvec3(50.)), //Light
-		new Triangle(dvec3( 50.,81.6-9.,76.6),dvec3( 42.,81.6-9.,76.6),dvec3( 50.,81.6-9.,81.6), dvec3(1.0), dvec3(50.)), //Light
+		/*new Sphere(600., vec3( 50.,681.6-.27,81.6), vec3(1.0),vec3(12.)) //Light */
+		/*new Sphere(5.,vec3( 50.,81.6-6.,81.6), vec3(1.0), vec3(12.)), //Light*/
+		/* Sphere(20.,vec3( 50.,51.6-6.,81.6), vec3(1.0), vec3(12.)) //Light */
+		new Triangle(vec3( 50.,81.6-9.,81.6),vec3( 42.,81.6-9.,76.6),vec3( 42.,81.6-9.,81.6), vec3(1.0), vec3(50.)), //Light
+		new Triangle(vec3( 50.,81.6-9.,76.6),vec3( 42.,81.6-9.,76.6),vec3( 50.,81.6-9.,81.6), vec3(1.0), vec3(50.)), //Light
 	});
 
 	pixel pixels[pixels_height][pixels_width];
 
 	// std::tan() is not constexpr
-	constexpr double FOV = 75.0;
-	const double transVal =  std::tan(glm::radians(FOV/2.0));
+	constexpr float FOV = 75.0;
+	const float transVal =  std::tan(glm::radians(FOV/2.0));
 
-	constexpr double aspect_ratio =
-		double(pixels_width)/double(pixels_height);
+	constexpr float aspect_ratio =
+		float(pixels_width)/float(pixels_height);
 
-	/* // Ray ray(dvec3(0.0), dvec3(0.0, 0.0, 1.0)); */
-	dvec3 Left =    dvec3( 1.0,  0.0,  0.0);
-	dvec3 Up   =    dvec3( 0.0,  1.0,  0.0);
-	dvec3 D    =    dvec3( 0.0,  0.0,  1.0);
+	/* // Ray ray(vec3(0.0), vec3(0.0, 0.0, 1.0)); */
+	vec3 Left =    vec3( 1.0,  0.0,  0.0);
+	vec3 Up   =    vec3( 0.0,  1.0,  0.0);
+	vec3 D    =    vec3( 0.0,  0.0,  1.0);
 
 	D          =  rotateY(D,      glm::radians(180.0));
 	Up         =  rotateY(Up,     glm::radians(180.0));
 	Left       =  rotateY(Left,   glm::radians(180.0));
 	
-	Ray ray(dvec3(50.0, 50.0, 150.0), D);
+	Ray ray(vec3(50.0, 50.0, 150.0), D);
    
 	auto start = high_resolution_clock::now();
 	
@@ -101,13 +101,13 @@ int main(int argc, char** argv) {
 		// does not mess with stdout so there
 		// is no flushing text like log info
 		fprintf(stderr, "\rRendering... %.1f%%",100.0 *
-				static_cast<double>((i+1)*pixels_width)/
-				static_cast<double>(pixels_width*pixels_height));
+				static_cast<float>((i+1)*pixels_width)/
+				static_cast<float>(pixels_width*pixels_height));
 	   
 		for(uint32_t j = 0; j < pixels_width; ++j){
 			dvec2 uv = dvec2(
-						  double(j)/double(pixels_width-1),
-					1.0 - double(i)/double(pixels_height-1));
+						  float(j)/float(pixels_width-1),
+					1.0 - float(i)/float(pixels_height-1));
 
 			uv = uv * 2.0*transVal - transVal;
 			uv.x *= aspect_ratio;
@@ -118,19 +118,19 @@ int main(int argc, char** argv) {
 			ray.D = normalize(ray.D);
 			
 
-			dvec3 col = dvec3(0.0);
+			vec3 col = vec3(0.0);
 			
 			for(uint32_t k = 0; k < numSamples; ++k)
 				col += tracer.radiance(ray, &cornellBox);
-			/*double t;*/
-			/*if((t=Triangle(dvec3( 50.,81.6-9.,-81.6),dvec3( 51.,81.6-9.,81.6),dvec3( 42.,81.6-9.,81.6), dvec3(1.0), dvec3(12.)).intersect(ray)) < MAX_T){*/
-			/*	col = dvec3(0.0, 1.0, 0.0);*/
+			/*float t;*/
+			/*if((t=Triangle(vec3( 50.,81.6-9.,-81.6),vec3( 51.,81.6-9.,81.6),vec3( 42.,81.6-9.,81.6), vec3(1.0), vec3(12.)).intersect(ray)) < MAX_T){*/
+			/*	col = vec3(0.0, 1.0, 0.0);*/
 			/*	printf("%lf\n", t); */
 			/*} else {*/
-			/*	col = dvec3(0.0);*/
+			/*	col = vec3(0.0);*/
 			/*}*/
 			//
-			col *= double(1.0/numSamples);
+			col *= float(1.0/numSamples);
 			
 			pixels[i][j] = toPixel(col);
 		}
@@ -140,7 +140,7 @@ int main(int argc, char** argv) {
 
 	auto end = high_resolution_clock::now();
 	auto time = (end - start)/std::chrono::milliseconds(1);
-	printf("time: %.3lf\n", static_cast<double>(time)*1e-3);
+	printf("time: %.3f\n", static_cast<float>(time)*1e-3);
 	
 	stbi_write_png("outimage.png",
 	               pixels_width, pixels_height, 3,
